IO/DataReader: split to class PinnedBuffer
Only very few callers really need the (stateful) DataReader features;
all they need is obtain the pointers to the pinned buffer.
